requirements.txt
alembic==1.13.1
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
dominate==2.9.1
email_validator==2.2.0
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==1.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Uploads @ git+https://github.com/maxcountryman/flask-uploads@f66d7dc93e684fa0a3a4350a38e41ae00483a796
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
jinja-markdown==1.210911
Jinja2==3.1.4
Mako==1.3.5
Markdown==3.6
MarkupSafe==2.1.5
numpy==2.0.0
pandas==2.2.2
Pygments==2.18.0
pymdown-extensions==10.8.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
setuptools==69.5.1
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
tzdata==2024.1
visitor==0.1.3
Werkzeug==3.0.3
wheel==0.43.0
WTForms==3.1.2
